refactor: Board, BoardInitializer 패키지 변경
